+++ src/ui/driver/drivercommon.h
@@ -30,13 +30,13 @@ int bufferSize();
 quint32 confIoConfOff();
 
 quint32 logAppHeaderSize();
-quint32 logAppSize(quint32 pathLen);
+quint32 logAppSize(quint16 pathLen);
 
 quint32 logConnHeaderSize(bool isIPv6 = false);
-quint32 logConnSize(quint32 pathLen, bool isIPv6 = false);
+quint32 logConnSize(quint16 pathLen, bool isIPv6 = false);
 
 quint32 logProcNewHeaderSize();
-quint32 logProcNewSize(quint32 pathLen);
+quint32 logProcNewSize(quint16 pathLen);
 
 quint32 logStatHeaderSize();
 quint32 logStatTrafSize(quint16 procCount);
@@ -46,14 +46,14 @@ quint32 logTimeSize();
 
 quint8 logType(const char *input);
 
-void logAppHeaderWrite(char *output, bool blocked, quint32 pid, quint32 pathLen);
-void logAppHeaderRead(const char *input, int *blocked, quint32 *pid, quint32 *pathLen);
+void logAppHeaderWrite(char *output, bool blocked, quint32 pid, quint16 pathLen);
+void logAppHeaderRead(const char *input, int *blocked, quint32 *pid, quint16 *pathLen);
 
-void logConnHeaderWrite(char *output, PCFORT_CONF_META_CONN conn, quint32 pathLen);
-void logConnHeaderRead(const char *input, PFORT_CONF_META_CONN conn, quint32 *pathLen);
+void logConnHeaderWrite(char *output, PCFORT_CONF_META_CONN conn, quint16 pathLen);
+void logConnHeaderRead(const char *input, PFORT_CONF_META_CONN conn, quint16 *pathLen);
 
-void logProcNewHeaderWrite(char *output, quint32 pid, quint32 pathLen);
-void logProcNewHeaderRead(const char *input, quint32 *pid, quint32 *pathLen);
+void logProcNewHeaderWrite(char *output, quint32 pid, quint16 pathLen);
+void logProcNewHeaderRead(const char *input, quint32 *pid, quint16 *pathLen);
 
 void logStatTrafHeaderRead(const char *input, quint16 *procCount);
